Separating NotInstruction and NullCheckInstruction to separate files.
